bot/requirements.txt
python-telegram-bot
google-generativeai
openai
python-dotenv
